Refactor app context
