+++ nad_ch/cli.py
@@ -1,5 +1,5 @@
 from nad_ch.controllers.cli import cli
-from nad_ch.application_context import create_app_context
+from nad_ch.config import create_app_context
 
 
 def main():
