+++ nad_ch/config/__init__.py
@@ -1,7 +1,9 @@
 from .base import APP_ENV
 
 
-if APP_ENV == "dev_local" or APP_ENV == "test":
+if APP_ENV == "dev_local":
     from .development_local import *
 elif APP_ENV == "dev_remote":
     from .development_remote import *
+elif APP_ENV == "test":
+    from .test import *
