+++ nad_ch/infrastructure/task_queue.py
@@ -1,8 +1,12 @@
+import os
 from celery import Celery
 import geopandas as gpd
 from nad_ch.application.interfaces import TaskQueue
 from nad_ch.application.validation import get_feature_count
-from nad_ch.config import QUEUE_BROKER_URL, QUEUE_BACKEND_URL
+
+
+QUEUE_BROKER_URL = os.getenv("QUEUE_BROKER_URL")
+QUEUE_BACKEND_URL = os.getenv("QUEUE_BACKEND_URL")
 
 
 celery_app = Celery(
